move useConnectionStatus hook
