+++ packages/react/src/components/ConnectionState.tsx
@@ -1,8 +1,6 @@
-import { connectionStateObserver } from '@livekit/components-core';
 import type { Room } from 'livekit-client';
 import * as React from 'react';
-import { useEnsureRoom } from '../context';
-import { useObservableState } from '../hooks/internal/useObservableState';
+import { useConnectionState } from '../hooks';
 
 /** @public */
 export interface ConnectionStatusProps extends React.HTMLAttributes<HTMLDivElement> {
@@ -10,23 +8,6 @@ export interface ConnectionStatusProps extends React.HTMLAttributes<HTMLDivEleme
    * The room from which the connection status should be displayed.
    */
   room?: Room;
-}
-
-/**
- * The `useConnectionState` hook allows you to simply implement your own `ConnectionState` component.
- *
- * @example
- * ```tsx
- * const connectionState = useConnectionState(room);
- * ```
- * @public
- */
-export function useConnectionState(room?: Room) {
-  // passed room takes precedence, if not supplied get current room context
-  const r = useEnsureRoom(room);
-  const observable = React.useMemo(() => connectionStateObserver(r), [r]);
-  const connectionState = useObservableState(observable, r.state);
-  return connectionState;
 }
 
 /**
