+++ packages/react/src/hooks/index.ts
@@ -1,4 +1,5 @@
 export { useLiveKitRoom } from './useLiveKitRoom';
+export { useConnectionState } from './useConnectionStatus';
 export { useAudioPlayback } from './useAudioPlayback';
 export { useDataChannel } from './useDataChannel';
 export { useGridLayout } from './useGridLayout';
